requirements.txt
pygame==1.9.3
Pillow==5.0.0